refactor flask server
