+++ app/core/__init__.py
@@ -1,0 +1,1 @@
+# Package initialization for core module
